Fix crash in SSLabel dealloc when there are no registered observers.
The dealloc method attempts to remove the observers added in the initWithFrame method but if the initWithFrame has not been called such as when using a SSLabel in a view without a parent in a XIB then it crashes with an exception but it is not allowed to remove an observer than hasn't been added.

This fix does a simple check to see if there are any registered observers before attempting to remove them.